mods_panel create mods_panel_list_widget.py and move class ModListWidget to it
